+++ src/ansys/fluent/core/services/datamodel_se.py
@@ -36,6 +36,9 @@ from ansys.api.fluent.v0 import datamodel_se_pb2_grpc as DataModelGrpcModule
 from ansys.api.fluent.v0.variant_pb2 import Variant
 from ansys.fluent.core.data_model_cache import DataModelCache, NameKey
 from ansys.fluent.core.module_config import config
+from ansys.fluent.core.services._command_arguments_mixin import (
+    CommandArgumentsCleanupMixin,
+)
 from ansys.fluent.core.services.interceptors import (
     BatchInterceptor,
     ErrorStateInterceptor,
@@ -481,7 +484,7 @@ class SubscriptionList:
                 v.unsubscribe()
 
 
-class DatamodelService(StreamingService):
+class DatamodelService(CommandArgumentsCleanupMixin, StreamingService):
     """Pure Python wrapper of DatamodelServiceImpl."""
 
     def __init__(
@@ -503,56 +506,7 @@ class DatamodelService(StreamingService):
         self.file_transfer_service = file_transfer_service
         self.cache = DataModelCache() if config.datamodel_use_state_cache else None
         self.version = version
-        self._command_arguments: set[tuple[str, str, str, str]] = set()
-        self._command_arguments_lock = RLock()
-        self._shutdown_cleanup_started = False
-
-    @staticmethod
-    def _command_arguments_key(
-        rules: str, path: str, command: str, commandid: str
-    ) -> tuple[str, str, str, str]:
-        return (rules, path, command, commandid)
-
-    def register_command_arguments(
-        self, rules: str, path: str, command: str, commandid: str
-    ) -> None:
-        """Register command arguments for explicit cleanup during shutdown."""
-        with self._command_arguments_lock:
-            if self._shutdown_cleanup_started:
-                return
-            self._command_arguments.add(
-                self._command_arguments_key(rules, path, command, commandid)
-            )
-
-    def release_command_arguments(
-        self, rules: str, path: str, command: str, commandid: str
-    ) -> None:
-        """Release command arguments when their Python wrapper is deleted."""
-        key = self._command_arguments_key(rules, path, command, commandid)
-        with self._command_arguments_lock:
-            was_registered = key in self._command_arguments
-            self._command_arguments.discard(key)
-            shutdown_cleanup_started = self._shutdown_cleanup_started
-
-        if was_registered and not shutdown_cleanup_started:
-            self._delete_command_arguments_rpc(rules, path, command, commandid)
-
-    def delete_all_command_arguments(self) -> None:
-        """Delete all tracked command arguments as part of shutdown finalization."""
-        with self._command_arguments_lock:
-            self._shutdown_cleanup_started = True
-            command_arguments = list(self._command_arguments)
-            self._command_arguments.clear()
-
-        for rules, path, command, commandid in command_arguments:
-            try:
-                self._delete_command_arguments_rpc(rules, path, command, commandid)
-            except Exception as exc:
-                logger.info(
-                    "delete_all_command_arguments %s: %s",
-                    type(exc).__name__,
-                    exc,
-                )
+        self._init_command_arguments_cleanup()
 
     def _delete_command_arguments_rpc(
         self, rules: str, path: str, command: str, commandid: str
@@ -741,11 +695,7 @@ class DatamodelService(StreamingService):
         self, rules: str, path: str, command: str, commandid: str
     ) -> None:
         """Delete command arguments."""
-        with self._command_arguments_lock:
-            self._command_arguments.discard(
-                self._command_arguments_key(rules, path, command, commandid)
-            )
-        self._delete_command_arguments_rpc(rules, path, command, commandid)
+        return super().delete_command_arguments(rules, path, command, commandid)
 
     def get_static_info(self, rules: str) -> dict[str, Any]:
         """Get static info."""
